requirements.txt
streamlit>=1.30.0
img2pdf>=0.5.1
Pillow>=10.0.0
